+++ src/main/java/mikera/matrixx/impl/SparseRowMatrix.java
@@ -130,7 +130,7 @@ public class SparseRowMatrix extends ASparseRCMatrix implements ISparse, IFastRo
 //	}
 
 	@Override
-	protected int componentCount() {
+	protected long componentCount() {
 		return rows;
 	}
 
@@ -335,8 +335,8 @@ public class SparseRowMatrix extends ASparseRCMatrix implements ISparse, IFastRo
 
 	@Override
 	public AMatrix multiplyCopy(double a) {
-		int n=componentCount();
-		AVector[] ndata=new AVector[n];
+		long n=componentCount();
+		AVector[] ndata=new AVector[(int)n];
 		for (int i = 0; i < n; ++i) {
             AVector v = unsafeGetVec(i);
             if (v != null)
